# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API docs for Auto Trade",
        "title": "TOC MACHINE TRADING",
        "contact": {},
        "version": "2.5.0"
    },
    "paths": {
...
        "/v1/order": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "manualInsertFutureOrder",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.manualInsertFutureOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/order/account/balance": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "getAccountBalance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.accountSummary"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/order/all": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "getAllOrder",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.allOrder"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/order/balance": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "getAllTradeBalance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.tradeBalance"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/order/balance/future/last": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "getLastFutureTradeBalance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.FutureTradeBalance"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/order/balance/stock/last": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "getLastStockTradeBalance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.StockTradeBalance"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/order/date/{tradeday}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "getAllOrderByTradeDay",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tradeday",
                        "name": "tradeday",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.futureOrders"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "updateTradeBalanceByTradeDay",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tradeday",
                        "name": "tradeday",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.futureOrders"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/order/day-trade/forward": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "calculateForwardDayTradeBalance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "buy_price",
                        "name": "buy_price",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "buy_quantity",
                        "name": "buy_quantity",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sell_price",
                        "name": "sell_price",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sell_quantity",
                        "name": "sell_quantity",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.dayTradeResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/order/day-trade/reverse": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "calculateReverseDayTradeBalance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sell_first_price",
                        "name": "sell_first_price",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sell_first_quantity",
                        "name": "sell_first_quantity",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "buy_later_price",
                        "name": "buy_later_price",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "buy_later_quantity",
                        "name": "buy_later_quantity",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.dayTradeResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/order/future/{order-id}": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "moveFutureOrderToLatestTradeDay",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order-id",
                        "name": "order-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/order/stock/{order-id}": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "moveStockOrderToLatestTradeDay",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order-id",
                        "name": "order-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.AccountBalance": {
            "type": "object",
            "properties": {
                "available_margin": {
                    "type": "number"
                },
                "balance": {
                    "type": "number"
                },
                "bank_id": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "risk_indicator": {
                    "type": "number"
                },
                "today_margin": {
                    "type": "number"
                },
                "yesterday_margin": {
                    "type": "number"
                }
            }
        },
        "entity.Future": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "delivery_date": {
                    "type": "string"
                },
                "delivery_month": {
                    "type": "string"
                },
                "limit_down": {
                    "type": "number"
                },
                "limit_up": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "reference": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "underlying_kind": {
                    "type": "string"
                },
                "unit": {
                    "type": "integer"
                },
                "update_date": {
                    "type": "string"
                }
            }
        },
        "entity.FutureOrder": {
            "type": "object",
            "properties": {
                "base_order": {
                    "$ref": "#/definitions/entity.OrderDetail"
                },
                "code": {
                    "type": "string"
                },
                "future": {
                    "$ref": "#/definitions/entity.Future"
                },
                "position": {
                    "type": "integer"
                }
            }
        },
        "entity.FutureTradeBalance": {
            "type": "object",
            "properties": {
                "forward": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "reverse": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "trade_count": {
                    "type": "integer"
                },
                "trade_day": {
                    "type": "string"
                }
            }
        },
...
        "entity.OrderAction": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "ActionNone",
                "ActionBuy",
                "ActionSell"
            ]
        },
        "entity.OrderDetail": {
            "type": "object",
            "properties": {
                "action": {
                    "$ref": "#/definitions/entity.OrderAction"
                },
                "order_id": {
                    "type": "string"
                },
                "order_time": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/entity.OrderStatus"
                }
            }
        },
...
        "entity.StockOrder": {
            "type": "object",
            "properties": {
                "base_order": {
                    "$ref": "#/definitions/entity.OrderDetail"
                },
                "lot": {
                    "type": "integer"
                },
                "share": {
                    "type": "integer"
                },
                "stock": {
                    "$ref": "#/definitions/entity.Stock"
                },
                "stock_num": {
                    "type": "string"
                }
            }
        },
...
        "entity.StockTradeBalance": {
            "type": "object",
            "properties": {
                "discount": {
                    "type": "integer"
                },
                "forward": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "original_balance": {
                    "type": "integer"
                },
                "reverse": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "trade_count": {
                    "type": "integer"
                },
                "trade_day": {
                    "type": "string"
                }
            }
        },
...
        "resp.Response": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string"
                }
            }
        },
        "v1.accountSummary": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.AccountBalance"
                    }
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "v1.allOrder": {
            "type": "object",
            "properties": {
                "future": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FutureOrder"
                    }
                },
                "stock": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.StockOrder"
                    }
                }
            }
        },
        "v1.dayTradeResult": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                }
            }
        },
        "v1.futureOrders": {
            "type": "object",
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FutureOrder"
                    }
                }
            }
        },
...
        "v1.manualInsertFutureOrderRequest": {
            "type": "object",
            "required": [
                "action",
                "code",
                "order_time",
                "price",
                "quantity"
            ],
            "properties": {
                "action": {
                    "$ref": "#/definitions/entity.OrderAction"
                },
                "code": {
                    "type": "string"
                },
                "order_time": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "v1.tradeBalance": {
            "type": "object",
            "properties": {
                "future": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FutureTradeBalance"
                    }
                },
                "stock": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.StockTradeBalance"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}